Setup project and initial sources
This is the initial commit of the project. Setup the file structure, add
a README file and the initial source files. The tool as it is
understands the `ips_list.yml` and `src_files.yml` files, and can
clone the dependency IPs into the `.landa` directory. Nothing happens
after that yet. The tool already reads the clone packages' config files,
but does not resolve their dependencies, nor does it provide any useful
list of sources. This is for future me to code.