Add dependency resolver stub
Add the `resolver` module which contains the `DependencyResolver`. This
will later contain the dependency resolution table and iteratively
settle on specific versions. The resolver shall be constructed from an
optional lock file, and will produce a new lock file upon destruction.

Also add a stub dependency internalization mechanism for the `Session`.
This allows dependencies from packages to be mapped to a unique ID that
can be used for further investigation and resolution.