Impl first git db checkout
Implement the creation and updating of the git database repositories.
These are held in a separate directory, usually `~/.landa/git/db` and
reflect each dependency that is in use. We use bare repositories here
to have a local mirror that we can query quickly about tags, branches,
commits, and individual files for use during dependency resolution.

Note that the git commands are currently executed using the standard
`std::process::Command`. Later we'll want to move everything into the
`git` module and use the `spawn_async` function from `tokio_process`.