Migrate to tokio-process
Perform the git operations using `tokio-process`. The actual processes
seem to execute, but the future never resolves. This is due to the fact
that the tokio core is currently inside the `Session` where it cannot be
accessed. Rather, there should be a `SessionIo` that wraps around
`Session` and keeps a tokio core that can then be executed.