Gather versions for git deps
Implement the gathering of the versions, refs, and revisions for git
dependencies. Also clean up the code in `sess.rs` a bit. Also extend the
`Git::list_refs()` function to not directly return the hashes, but
rather resolve them to actual commits. This handles the case of
annotated tags properly.