Add deps to resolve table, prepare constraints
After the dependencies have been added to the table of the resolver,
initialize all dependencies with state `Open`. Also add a representation
for a dependency constraint which can be derived from a manifest's
dependency specification. Add the first steps to the `mark()` function
of the resolver which will go through each dependency's list of versions
and impose the constraints of other packages.