Fix empty dependency repositories
Change the way empty dependency repositories are handled. Make the call
to `git show-ref` unchecked such that it does not cause an error if no
refs are found. This is the case in empty repositories. Then also adjust
the way refs and revs are collected by the session: Rather than doing
both at the same time, make them depend on each other. Only collect the
revs if there were any refs in the repo.

This fixes #2.