Fix creation of checkouts
The last commit broke the creation of checkouts. `--separate-git-dir`
causes the database repository to keep a list of work trees. This
becomes a problem as soon as we delete such a work tree, rendering the
database useless. Use `git archive` to create a tarball of the files,
then switch to the intended checkout directory and unpack. Works well.