Finish initial recursive resolution
The dependency resolver did not properly recover from cases where a
picked dependency version would become invalid. This is now fixed. The
whole resolver code is quite convoluted and messy. It should be
rewritten against a clean `DependencyProvider` trait which we can then
use to create mockups for unit tests. But the initial version should be
up and running.