Only run resolution on outdated lockfiles
Rather than running dependency resolution all the time, check whether
the manifest has been modified after the lock file was created. If that
is the case, perform dependency resolution. Otherwise just reuse the
contents of the lock file. This makes resolution the rare case, only
happening upon changes to the manifest. Great speedup for the tool.