Add sources subcommand, fix path deps
Add the `sources` subcommand which emits a source file manifest in YAML
format to stdout. This allows other tools to ingest the sources and
operate on them appropriately.

Fix an issue with path dependencies where their manifest would not be
read properly. This was due to the paths in the manifest being left
relative, rather than properly anchoring them to wherever the checkout
of the dependency is.