Add target spec to source group
Implement target specifications. These are a form of boolean expressions
to declare functions on the presence of certain target flags. Create a
parser for them and add a `target` field to the source file group. This
now allows the user to specify e.g. `target: all(fpga, not(xilinx))`
together with their source files.