Add ability to override dependencies
Add a `overrides` field to the configuration files. These allow
dependencies to be overridden locally. I expect this to be used mainly
for local working copies overridden as path dependencies. Not happy with
the way this is currently handled. Maybe the session should properly
process the manifests it reads from disk and apply the overrides as
appropriate.