Add config cmd, plugins in config
Add the `bender config` command to dump the tool configuration as JSON.
By default only the parsed options will be dumped, discarding any of the
values that were unknown. This will change in the future.

Add the `plugins` section to the configuration filesm, which allows for
package dependencies to be declared per-machine rather than per-repo.
It's quite a hack right now and must be cleaned up further.

Also have the source groups track from which package they came from.