Improve unsatisfiable constraint error messages
Keep track of which package causes which constraint. If a dependency
cannot satisfy all constraints, print a list of constraints and which
package imposes them. Also make the error message clearer that indicates
if a package cannot meet a constraint at all. This was usually caused by
having multiple sources for the same package, where some of the sources
could satisfy a constraint while others could not.