Fix path dep resolution issue
Fix an issue in the dependency resolver where a path dependency would
not properly trigger a second resolution iteration to resolve its own
dependencies. This would result in a panic due to those dependencies not
having a version picked.

To fix this, add the missing `any_changes = true` statement to the
segment of code executed whenever a "constrained" dependency is changed
to the "picked" state.