Add `workspace` section
Add the `workspace` section to the package manifest which allows for
certain aspects of the workspace to be configured. Move the
`package_links` setting in there. Add a `checkout_dir` option to the
workspace which will cause Bender to keep a human-editable checkout of
each dependency at hand and not interfere with it once checked out.
Further tooling will be necessary to guide the user through upgrading
dependencies.