vendor: prioritize file mappings over directories
File mappings are more specific and
should be preferred over directory mappings.

As an example we have two mappings:
a/ -> c/
b/file -> c/file

Any change in c/ should be then result in a patch.
If the change is in c/file we must make sure it is linked back
to b/file, not erroneously to a/, the sorting does this.

Similarly when initializing we first need to copy the a/ dir to c/
and only then copy b/file to c/file.