git: prevent `git tag` opening an editor and hanging
For me, the `git tag bender-tmp-<commit> <commit> --force` opens a
nano process, which then causes the `bender checkout` to just hang
and so it just freezes. This is because I have `tag.gpgSign = true`
set in my git config.

To resolve this, pass the `--no-sign` flag so that we don't open
an editor.

Signed-off-by: julia <git.ts@trainwit.ch>