treewide: Add heuristic filtering of unused sources
We build a tree of deps from the manifests and then scan the paths to
filter out unused modules from the manifests. The filtering is heuristic
because it relies on regex pattern matching between the manifests of
packages and the modules. Currently supports only verilog/system
verilog. Feature is experimental.